+++ gnovm/stdlibs/chain/address.gno
@@ -1,4 +1,4 @@
-package std
+package chain
 
 import (
 	"crypto/bech32"
