+++ gnovm/stdlibs/chain/banker/coins.gno
@@ -1,4 +1,4 @@
-package std
+package banker
 
 import (
 	"math/overflow"
